glab: fix non-bash shell completion
Because `glab` expects the `-s`/`--shell` option, the formula
mistakenly installed the same bash completion for both zsh and fish
as well.

Revision bump to "push" fix to users.

Finally, I simplified the code a bit because AFAICT the current method
of manipulating the `SHELL` environment variable is either vestigial
or it actually never had the intended effect.

Co-authored-by: Rylan Polster <rslpolster@gmail.com>